move api functions from needed by other api to service
