+++ src/api/upload.ts
@@ -1,52 +1,12 @@
-import { getZipFileStream } from './../services/fs';
 import { getElectronFile } from './../services/fs';
-import path from 'path';
-import StreamZip from 'node-stream-zip';
 import { uploadStatusStore } from '../stores/upload.store';
-import { ElectronFile, FILE_PATH_KEYS, FILE_PATH_TYPE } from '../types';
+import { ElectronFile, FILE_PATH_TYPE } from '../types';
 import { logError } from '../utils/logging';
 import { ipcRenderer } from 'electron';
-import { getSavedFilePaths } from '../services/upload';
-
-async function getZipEntryAsElectronFile(
-    zip: StreamZip.StreamZipAsync,
-    entry: StreamZip.ZipEntry
-): Promise<ElectronFile> {
-    return {
-        path: entry.name,
-        name: path.basename(entry.name),
-        size: entry.size,
-        lastModified: entry.time,
-        stream: async () => {
-            return await getZipFileStream(zip, entry.name);
-        },
-        blob: async () => {
-            const buffer = await zip.entryData(entry.name);
-            return new Blob([new Uint8Array(buffer)]);
-        },
-        arrayBuffer: async () => {
-            const buffer = await zip.entryData(entry.name);
-            return new Uint8Array(buffer);
-        },
-    };
-}
-
-export const setToUploadFiles = (type: FILE_PATH_TYPE, filePaths: string[]) => {
-    const key = FILE_PATH_KEYS[type];
-    if (filePaths) {
-        uploadStatusStore.set(key, filePaths);
-    } else {
-        uploadStatusStore.delete(key);
-    }
-};
-
-export const setToUploadCollection = (collectionName: string) => {
-    if (collectionName) {
-        uploadStatusStore.set('collectionName', collectionName);
-    } else {
-        uploadStatusStore.delete('collectionName');
-    }
-};
+import {
+    getElectronFilesFromGoogleZip,
+    getSavedFilePaths,
+} from '../services/upload';
 
 export const getPendingUploads = async () => {
     const filePaths = getSavedFilePaths(FILE_PATH_TYPE.FILES);
@@ -71,24 +31,6 @@ export const getPendingUploads = async () => {
         collectionName,
         type,
     };
-};
-
-export const getElectronFilesFromGoogleZip = async (filePath: string) => {
-    const zip = new StreamZip.async({
-        file: filePath,
-    });
-
-    const entries = await zip.entries();
-    const files: ElectronFile[] = [];
-
-    for (const entry of Object.values(entries)) {
-        const basename = path.basename(entry.name);
-        if (entry.isFile && basename.length > 0 && basename[0] !== '.') {
-            files.push(await getZipEntryAsElectronFile(zip, entry));
-        }
-    }
-
-    return files;
 };
 
 export const showUploadDirsDialog = async () => {
@@ -134,3 +76,9 @@ export const showUploadZipDialog = async () => {
         logError(e, 'error while selecting zips');
     }
 };
+
+export {
+    setToUploadFiles,
+    getElectronFilesFromGoogleZip,
+    setToUploadCollection,
+} from '../services/upload';
